[design/docs] TestMainBanner 컴포넌트 파일명 -> MainBanner로 변경하여 메인 홈 import 수정 / OverlayFilter를 슬라이드 전체 영역이 아닌 개별 영역 안으로 필터 적용되도록 위치 수정, filter 배경값 수치 살짝 조정함
