+++ src/pages/home/main/MainHome.jsx
@@ -20,54 +20,57 @@ import { getAiringToday, getMovie, getTopRated, getTVseries } from '../../../sto
 
 import { getAnimations, getDarkTheaterReleases, getDramaTvs } from '../../../store/modules/getThunkThree';
 import MainCommonList from '../../../components/home/common/MainCommonList';
+
 // import { LiveContainer } from '../live/style';
 // import { Container } from '../../../components/home/viboOnly/only1/style';
 
 const liveTitles = [{ id: '1', title: '실시간 인기 LIVE' }];
 
 const MainHome = () => {
-    const dispatch = useDispatch();
-    useEffect(() => {
-        dispatch(getMovie());
-        dispatch(getTVseries());
-        dispatch(getAiringToday());
-        dispatch(getTopRated());
-        dispatch(getAnimations());
-        dispatch(getDarkTheaterReleases());
-        dispatch(getDramaTvs());
-    }, []);
-    return (
-        <MainContent>
-            <MainBanner />
-            <MainContainer>
-                <div>
-                    {liveTitles.map((section) => (
-                        <LiveList key={section.id} title={section.title} />
-                    ))}
-                </div>
-                {/* <WatchingList /> */}
-                <WeeklyList />
-                <RateList />
-                <GenreNav />
-            </MainContainer>
-            <AdBanner />
-            <MainContainer>
-                {/* <CommonList type="series" />
+  const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(getMovie());
+    dispatch(getTVseries());
+    dispatch(getAiringToday());
+    dispatch(getTopRated());
+    dispatch(getAnimations());
+    dispatch(getDarkTheaterReleases());
+    dispatch(getDramaTvs());
+  }, []);
+  return (
+    <MainContent>
+      {/* <MainBanner /> */}
+      {/* <TestMainBanner /> */}
+      <MainBanner />
+      <MainContainer>
+        <div>
+          {liveTitles.map((section) => (
+            <LiveList key={section.id} title={section.title} />
+          ))}
+        </div>
+        {/* <WatchingList /> */}
+        <WeeklyList />
+        <RateList />
+        <GenreNav />
+      </MainContainer>
+      <AdBanner />
+      <MainContainer>
+        {/* <CommonList type="series" />
         <CommonList type="movie" />
         <CommonList type="series" /> */}
-                <MainCommonList fetchFunction={getAnimations} stateSelector={(state) => state.filterR.animations} />
-                <MainCommonList
-                    fetchFunction={getDarkTheaterReleases}
-                    stateSelector={(state) => state.filterR.darkTheaterReleases}
-                />
-                {/* <CommonList fetchFunction={getDramaTvs} stateSelector={(state) => state.filterR.DramaTvs} /> */}
-                {/* 헐 대소문자.. */}
-                <MainCommonList fetchFunction={getDramaTvs} stateSelector={(state) => state.filterR.dramaTvs} />
-                <ViboOnly />
-            </MainContainer>
-            <MenuSection />
-        </MainContent>
-    );
+        <MainCommonList fetchFunction={getAnimations} stateSelector={(state) => state.filterR.animations} />
+        <MainCommonList
+          fetchFunction={getDarkTheaterReleases}
+          stateSelector={(state) => state.filterR.darkTheaterReleases}
+        />
+        {/* <CommonList fetchFunction={getDramaTvs} stateSelector={(state) => state.filterR.DramaTvs} /> */}
+        {/* 헐 대소문자.. */}
+        <MainCommonList fetchFunction={getDramaTvs} stateSelector={(state) => state.filterR.dramaTvs} />
+        <ViboOnly />
+      </MainContainer>
+      <MenuSection />
+    </MainContent>
+  );
 };
 
 export default MainHome;
